fixed major bug
